Major refactor: Single scene and assembly
The previous version had gotten way more complicated than what it needed to be. This repo should be just a couple of boilerplate classes and the very basic scene structure so that we don't have to re-type those in every jam. Instead it ended up involving excess software engineering. We don't need another assembly (as we don't have that much code that assemblies would be beneficial), a state machine for scenes (`IScene` interface). For jam games, a single scene and a few gameobjects are more than enough.